+++ app/src/main/java/x/x/memlists/feature/lists/ListEditorScreen.kt
@@ -82,7 +82,7 @@ fun ListEditorScreen(
                         Text(
                             text = it,
                             color = palette.clText,
-                            fontSize = UiTokens.fsSmall
+                            fontSize = UiTokens.fsNormal
                         )
                     }
                     PrimaryActionButton(